Add classes, tests for reading/writing JSON
These are in squidlib-extra and complement the existing SquidStorage
class there, since SquidStorage has a requirement to use Preferences
from LibGDX, and these classes are more well-suited to data drawn from
files that are part of the game. DataConverter can convert various
classes that are common in SquidLib to special text that is mostly like
JSON, and can technically be edited by hand. For making hand-editable
code, though, you would find it easier to write your own serializer; the
serializer for RegExodus Pattern objects is the first in
DataConverter.initialize's code, and it's a good example for how you
could write your own with just two methods. There's also DataCompressor,
which is just like DataConverter but compresses/obfuscates the text it
produces and can read back that compressed text. DataCompressor has some
limits on what types of output and input it can use, due to the
compression needing to operate on Strings and not, say, arbitrary-length
InputStreams.

And there's a test too! It's meant for @SquidPony to take a look at and
see if this fits his use case. It loads a large amount of data that I
pasted from one of my public domain spreadsheets of game design stuff.
It seems correct, which was my main concern with it, and the code for
JSON handling seems pretty clean.