Optimizations for heavy usage of colors
A major performance problem was found in the last commit, but the issue
seems to go back much further. In any code that creates an extremely
large amount of Color objects and caches them with an IColorCenter,
performance slows to a crawl and stays slow even if the cache is
cleared. The issue does not appear if the number of requested colors is
fairly small (256 Color objects are no problem), but when running the
tests in HashVisualizer and every pixel on a 512x512 grid has its own
color, the speed is miserable. The issue is still present, somewhat, in
the current commit, but the many-colored hashing/RNG tests now are twice
as fast as before, and there's a lot less GC required as well for large
grids (we use a float[][] instead of a Color[][] in SquidPanel, which
helps).

There may be some changes here that break code if it assumes different
things about the generic type of colors used in an IColorCenter, but I
tried to reduce the impact that was present earlier in the course of
this commit and it might not affect subclasses at all. @smelc , this I
think could affect you, but there might be nothing to change.

I'm not sure how well IFilters work with the new code for more-optimized
float-based color specification.