Changed how OrderedMap is serialized; helps edits
Before, OrderedMap used the maybe-slightly-faster way of dumping the
keys to a newly-constructed OrderedSet and the values to a
newly-constructed ArrayList, and using those as collection arguments to
OrderedMap's constructor when deserializing. That had the unfortunate
issue of all keys being in one chunk, followed by all values, with no
easy way to tell which key corresponded to which value. This is changed
now so keys and values are in pairs, using Maker.makeOM() to deserialize
the special format of arguments, and that method now optionally allows a
float load factor. The new style may actually be more efficient by not
producing two data structures and using an array instead, and the method
it uses when deserializing is about the same.