Uh, wow. 10x speedup in HashVisualizer/SquidPanel
This should be great news for any games that use many small cells for
just their color. HashVisualizer now gets 30 frames per second,
sometimes more or less, displaying 512x512 individual cells with their
own individual colors, each randomly generated each frame. Before this
commit, it was getting about 3 frames per second, and a little before
that it was 1 or 2. The difference lies in using floats to represent
colors instead of their heavier LibGDX Color object form, and avoiding
re-filtering colors that were already filtered. Lookups in a cache of
Colors as objects in a HashMap seem to get slower and slower as the
cache grows, and this sidesteps the cache entirely by avoiding garbage
production.