Update build.gradle to download deps, wrapper
The build.gradle config has been updated to also build squidlib-extra
and to handle the assets needed for tests in both the display and -extra
modules. It downloads RegExodus and BlazingChain (a small dependency of
squidlib-extra) as part of the build now. The Gradle wrapper is version
3.1, which should be recent enough. All the tasks that run tests have
been checked and seem to work without error, at least here. There seems
to be some issue with the DijkstraMap pathfinding to the mouse cursor in
EverythingDemo though, unrelated to Gradle.