Move Gradle build files to zip to stop JitPack mess
You can unzip gradle_build.zip in the current folder (the "Extract Here"
option in applications like 7-Zip) to put the Gradle files in place to
use for building. If they are on Git and not in some archive, then
JitPack seems to assume we build purely with Gradle and ignores the
pom.xml files that actually do the correct setup for multiple modules.

I am not a fan of Gradle, and now I have yet another reason to give as
justification.