First draft of OptDijkstraMap; good speed
OptDijkstraMap is in squidlib-performance right now because it won't be
a visible class, probably; it should be be substituted in place of the
internals of CustomDijkstraMap once a few more parts fall into place.

The speed boost is nice, but not incredible; it takes 61 ms instead of
87 ms to find 599 paths in a 40x40 map with OptDijkstraMap instead of
CustomDijkstraMap. Still, it should be welcome.