Fix various bugs in OptDijkstraMap
These changes shouldn't affect speed at all, but do affect correctness.
The constant used as a poor-man's ceil() function was too close to 1.0
and was, due to floating-point math, being treated as 1.0; now we use
the largest possible float less than 1 instead of trying for the largest
possible double less than 1, which should still give ample room for
costs that are close to, but not equal to, 0. The map was treated as
mostly-reversed during one check for diagonal connectivity, which caused
problems, and there was a simple mistake in setFresh() that caused the
least-optimal distance to be used instead of the most-optimal. All three
are fixed, it seems, and a demo is now present at the botom of
OptDijkstraMap.java if anyone wants to see what the distance map looks
like.