Fix Adjacency.isBlocked(), prep OptDijkstraMap
OptDijkstraMap now works with doubles instead of ints for distances,
which makes it more comparable to the other Dijkstra-like pathfinders.
It isn't slower than before except on tiny paths, which are only
slightly slower and this may just be because it needs to work with twice
as many bytes to allocate double arrays compared to int arrays. I'm not
sure, though; it's a micro-optimization area. Adjacency had an issue
where it checked the wrong direction for determining diagonal blockage,
so valid paths were sometimes considered higher-cost and invalid paths
may have been considered viable for some part of the pathfinding process
in CustomDijkstraMap and OptDijkstraMap.

ThinWallDemo has some visual changes so the font looks a bit nicer.
Pathfinding seems to work in it now, though I'm not 100% certain yet.