NextDijkstraMap, various improvements.
NextDijkstraMap is meant to be an implementation that can replace most
(currently not all) of the existing DijkstraMap while being
significantly faster (about 2-3x currently). It doesn't implement
cover-seeking pathfinding, which doesn't seem to work at all in
DijkstraMap anyway, nor does it implement anything that uses an FOVCache
(not GWT-compatible, and not enough improvement on speed). Everything
else is there, including normal scanning and pathfinding, flee paths
(untested), large-size variants of the earlier methods (also untested,
but simpler than before), and technique/attack paths (still untested,
and without cover-seeking attack paths).

Other improvements include a way to use the FOV class via static methods
and reusing an existing double[][] for lights, some key additions to
Adjacency that are used as part of a fix in OptDijkstraMap (so soon,
CustomDijkstraMap), and a simple isEmpty() method in IntVLA.