Minor tweak to shader for distance field
Since we only set the uniform through TextCellFactory.configureShader()
, we can do a division in there (once per render, usually) and so avoid
that division, as well as a temporary value, being performed in the
fragment shader for every pixel. The numbers used are also tweaked
slightly to emphasize a more crisp-and-sharp appearance, which may also
help legibility.

The icons.png and atlas used in the assets now include basic UI
controls, so you can create a Skin using the icons atlas for most
controls (avoiding a texture swap between icons and UI elements). It
isn't the best Skin for a lot of uses, but... eh.