As expected, Wisp needed a small change
When hashing 32-bit or 64-bit values and generating a hash that has less
bits than a value, when values only differed in the most significant bit
or some number (n) of bits near it, then the hashes generated would
collide as if the values hashed were identical (ignoring the sign bit if
enough values were hashed). The problem got worse if more values needed
to be hashed, with about 1 additional bit of earlier input being
discarded/overridden for each additional value. This had an easy enough
solution, it seems; not left-shifting one of the values keeps the
most-significant bits while we can overwrite the least-significant bit
instead (changes in the LSB of the data propagate through most bits of
the result, while the same is not true for the MSBs, so the latter need
to be kept intact). This seems to have essentially no impact on
performance and improves quality, though it changes the hashes that Wisp
produces (it is still marked Beta, though, so this should be expected).