This should've been obvious... Dijkstra perf boost!
When finding one path, as opposed to scanning for the purpose of
generating a distance field, we don't need to keep scanning after we've
gone from a goal to the starting point -- that means we've found enough
for a path. With this simple optimization in place,
squidpony.squiddai.DijkstraMap outperforms gdx-ai's
IndexedAStarPathFinder (doing the same task, essentially, instead of
DijkstraMap doing more work before), at least on paths that are long and
complex enough for a simple A* heuristic to have some issues. The
changes here should be applied to OptDijkstraMap soon enough, and then
OptDijkstraMap should replace CustomDijkstraMap.