Another change to Wisp; helps avoid artifacts
The change primarily affects short arrays that are being hashed,
especially length-1 arrays, which had an issue where the high bits of
the hashed element didn't affect the low bits of the resulting hash very
much, but the reverse was true. Now we do some small and relatively
quick changes to the finalization to improve quality.