Changes to Ordered colls for better array keys
It isn't the default because array keys in Sets and Maps aren't common,
but you can pass CrossHash.generalHasher to an OrderedSet or OrderedMap
constructor to be able to hash multiple types of array keys. CrossHash's
IHasher values default to Wisp now, and Wisp is no longer Beta.

This may need some checks to ensure it works on GWT.