Remove most deprecated classes. #137 done.
Lots of documentation improvements, too. ProbabilityTable now stores a
separate internal collection of other ProbabilityTable items, and can
choose from one of those just like it can choose from one of its own
items based on weight. This is tested in ProbabilityTableTest in
squidlib-util. The deprecated classes from the display module are
removed, but TextPanel is staying (it wasn't deprecated, though it was
suggested for deletion, but it is in use in a few places). StarterKit
has a few improvements but still needs a demo. CrossHash has a hasher
that uses System.identityHashCode to hash items by identity and also
compares for equality by reference, which may be useful in some places
(it's already used by ProbabilityTable). We should be ready for 3.0.0-b7
now.