Close #42 . ArrayTools has array reversing methods.
SquidPanel now has at least one way to write a String in any of 8
directions. The behavior may be a little weird on diagonals because it's
meant to keep the bounds the same for right and left, as well as up and
down, so when the two are combined you get cases where the String may
never touch (xOffset,yOffset) because it's moving in a different way
across the same bounded space.

The reversal methods on ArrayTools should be handy in some cases, though
they turned out not to be needed in SquidPanel.