Back to SNAPSHOT. GreasedRegion can remove corners.
This corner-rounding is used unsuccessfully in BevelFOV, but BevelFOV
may be rewritten at some point to be more closely integrated with
GreasedRegion (allowing some things that are only in CoordPacker
currently to work in GreasedRegion). BevelFOV is not encouraged for any
use yet, since the results are identical to FOV with type SHADOW, but
it's slower.