Significant improvements to ObText
Docs are better, including a sample of the format. Raw strings allow
delimiters using a different syntax. Actually using the objects is
reasonable now because the way you get child elements is more robust
(returning a new Iterator over the child elements only, instead of
modifying the same iterator and dead-ending when the children run out).
And it's pretty small and convenient for a format, too!