Update to libGDX 1.9.6 (GWT 2.8.0). ObText stuff
You should update your GWT Eclipse plugin if you don't use Gradle, ahem,
@smelc , to 2.8.0. GWT 2.8.0 should be much better all around. The
build.gradle file also is updated, since it hadn't been in some time.

ObText now considers the character `#` to be the start of a line
comment, like `//` was already; `#!` was previously considered a comment
at the start of a line only, but this is more consistent. `/*block*/`
block comments are here too.