Starting work on StringConvert to help save/load
StringConvert should be a more standard way of translating objects to
and from Strings, and a few places already use it where non-standard
approaches were used before. ObText now implements Collection<String>,
so it should be a bit more useful. IntVLA has a way to store its
sequence of int as a CSV string and load from such a String as well.