Add TableSet, a Set-like that uses columns and rows
As a disclaimer, TableSet is at a very early stage and is marked Beta,
since it really needs testing. TableSet (and TableMap, which is coming
up next) may be useful for storing some common things in games, like the
case where you want a Coord as a key in a map that stores sparse data
but also want multiple layers in case there are duplicate entries at one
key. This could use a Coord and Integer as its C and R types for that,
though since it's Set-like it doesn't have additional values (it just
would be used to mark presence). OrderedSet and OrdererdMap are no
longer marked as Beta, since they've been shown to be stable now.

Also, woohoo, 1000 days commit streak as of today!