Various changes that help performance. Breaking?
I tried to avoid breaking any public APIs in SquidPanel and
TextCellFactory, but some of the protected fields changed in both. For
various reasons, SquidPanel performs a fair amount better by using a
char[][] instead of a String[][] for its contents, which matters a lot
for serialization.  I'm planning to allow saving and loading
possibly-layered maps with their color info intact, like a screenshot,
so this could be used to make something like a paint program or use such
a program's results. RexPaint already does this, but it has various
incompatible properties of its .xp format, from the gzip compression
being difficult on GWT to the lack of alpha or Unicode characters in a
file. I think there's an issue requesting this from way back... I am not
sure what subclassing of SquidPanel @smelc does, if any, but this might
take some minor tweaking to get working as long as nothing relies on
having multiple chars (or none, not even a space) in one cell.

A probably not-breaking change is the move of the static methods for
dealing with colors as floats in
`squidpony.squidgrid.gui.gdx.Filters.Utility` to
`squidpony.squidgrid.gui.gdx.SColor`, where they belong more.

There's also the always-nice getOrDefault method implemented in
OrderedMap. It's used in the swapping functionality to clean up some
code in TextCellFactory. StringKit has a bunch of additions too.