Switch to a more flexible StringConvert constructor
Now it takes CharSequence items, each of which is the name of a type.
The name could be that of a generic type, like to convert between
Strings and "OrderedMap<String, OrderedSet<String>>", you could use `new
StringConvert("OrderedMap", "String", "OrderedSet<String>")`, and this
handles nested types handily as long as the component types have been
registered (by constructing them). Because the type names (without
generics) match the results of `Class.getSimpleName()`, and because that
is one of the few Class methods present on GWT, we have a better way to
create or look up StringConverts for more-arbitrary data now.