Add fonts: Iosevka with Sans and Slab styles
Iosevka Slab in particular should fill a role no other font currently
does. Preview images are now linked in the documentation for stretchable
fonts in DefaultResources. As an aside, I'm using Iosevka now as my
IDE's main font and I love it, though this is partly attributable to the
customizations that building the font allows making it more
personalized, increasing my attachment to the font. I hope someone else
likes it too!