Add hashing methods to Wisp that do better on GWT
hashAlt and hash64Alt take float or double arrays, and are slower on
desktop, but should be much faster on GWT because they avoid
Double.doubleToLongBits(), which has a much more complex implementation
on GWT than on desktop. They also have better visual hashing
characteristics, with the original hash and hash64 showing streak-shaped
artifacts where x or y are low, but the new hashAlt and hash64Alt
showing none.