Bring in 6D Simplex, seamless 3D from Joise
This is in SeededNoise in squidmath. More kinds of noise should be
forthcoming, probably using the large lookup tables at the start of the
file. There are good demos of seamless noise in HashVisualizer, which
uses it to speed up rendering despite using a ton of trigonometry to
make it seamless.