Include some ways to pass noise functions around
This is used to make a simple way to layer different scales of noise
with Noise.layerOctaves1D() (up to 4D, plus 6D), or to modify it with a
"highlighted ridge" effect with Noise.ridged2D() (up to 4D, plus 6D).
What these return are classes that implement Noise.Noise2D (or Noise 3D,
4D, 6D, or sometimes 1D), which allows them to be passed to things that
want a certain dimensionality of noise, like SeededNoise.seamless3D() .
These can be previews for 2D and 3D in HashVisualizer.