Move ValueNoise to GlitchNoise, more stuff
There's still a file called ValueNoise, and it's actually closer to what
it should be given that name. GlitchNoise should be useful for, well,
making things look like the game is glitching up. Noise has the option
to scale the arguments you give it by using a Scaled"n"D instance, for n
in (1,2,3,4,6). Turbulent Noise should behave better.