Lots of additions to help port worldgen code
This is @zacharycarter 's mapgen project, which I helped on parts of,
and now I'm trying to bring it into SquidLib as either a demo or a
library class. Therre's now a frequency option in most kinds of Noise,
there's a way to lerp between two colors represented as floats (without
creating objects), and a few other nice things.