Tweaks to DetailedMap. Add MasonNoise.
MasonNoise is an attempt to speed up SeededNoise, though it isn't really
a success in that sense yet (speed is close to SeededNoise, but slightly
slower). It does have better appearance if you want to conceal patterns,
so there's that. Noise.Ridged"N"D classes now can setOctaves.
DetailedWorldMapDemo continues to improve in appearance and also
generation speed.