Hooray! Rivers work with zooming
This uses the new GreasedRegion.connect() (or rather, connect8way() )
method to merge separate points after zooming. Due to using
GreasedRegion.quasiRandomRegion() on the fringe of each river when we
zoom, then fusing that with the rivers, we have a nice wiggly shape to
the rivers, though they're a little thick upon zooming far enough.