2D array converters; boolean[] StringConvert change
Now Converters.convertBoolean will print the results much more concisely
(and readably) by using '1' for true, '0' for false, and sometimes 'n'
or 'N' for when a boxed Boolean was passed and it was null, or when all
or some part of a boolean[][] was null. This also is much simpler and
faster to parse, since it avoids handling substrings that can have
variable length ("true" having a different length than "false"). It
doesn't use delimiters for the 1D array because all possible values can
be represented with one char each, but Converters.convertArrayBoolean2D
, which handles boolean[][] values, uses ';' to delimit sub-arrays,
which allows the full boolean[][] to be jagged. All of the 2D array
StringConvert implementations allow jagged arrays in a similar way, and
except for Converters.convertArrayChar2D, they treat jagged arrays the
same as rectangular ones.