You can use SquidPanel to view a sub-area now
This is tested in EverythingDemo with a 270x75 map, showing a 90x25
sub-section at a time, centered on the player. This actually needed a
change to DungeonGenerator to produce a map as large as I requested (it
used CoordPacker for some things, but that is limited to a 256x256 map,
so I switched it to GreasedRegion and cleaned up some nearby stuff in
the process). SquidPanel and SquidLayers have changes to work with this.
There's also some stuff I was working on in DetailedWorldMapDemo, which
is still going nicely.