Allow float text size in TCF (match the rest of it)
This is a very minor breaking change if you assigned
TextCellFactory.width() or TextCellFactory.height() to an int, because
now those return floats. Internally, they were floats in most of
TextCellFactory, and were just cast to int in several places. If you
cast the results of width() or height() to int to keep your current
types, you shouldn't see any changes.

This commit also contains updates to several of the demos/examples in
the display module, and in squidlib-util, there's
FakeLanguageGen.SentenceForm to bundle up the sometimes-messy parameters
to FakeLanguageGen.sentence().

This set of work might not be 100% complete yet, but it builds and the
demos work just fine with some small changes to one (and the stretchable
fonts look quite a bit better, since the box drawing characters often
need the font to be stretched to take up an extra space that's more than
0 and less than 1 pixel in size).