Much nicer smooth scrolling in oversized maps
This uses a number of new techniques and may be overly complex for
EverythingDemo to show as an example, but it looks really good. We use a
Camera now in EverythingDemo, which needed some fiddling with a second
viewport to display the messages separately. Camera allows the scrolling
to be smooth and avoid gaps at the edges.