Many fixes; affects DungeonGenerator, Dijkstra
DungeonGenerator has changed what it will produce given the same seed a
few times recently, and it happened again now. First I found a better
way to do sampling of separated areas with less samples overlapping (it
should also be faster, but that isn't a main priority for dungeon gen).
Then, I found that grass wasn't being generated, which was an easy fix,
but also that doors weren't being generated either. The doors turned out
to be related to DijkstraMap not respecting impassable cells all the
time, which I think is fixed now.

EverythingDemo also has a nice fix to camera movement at the edges, so
it should be smoother throughout. This needed a few changes to other
parts of SquidPanel, but shouldn't affect existing code if it doesn't
use gridOffsetX or gridOffsetY to render only part of the map. Camera
movement in EverythingDemo is much less of a hack now, and doesn't rely
on animation timing as part of what determines positions any more,
wow...