Zoom anywhere in DetailedWorldMapDemo
This uses the mouse, with left click zooming in to the point you clicked
(or '+' zooming into the center of the screen) and right click zooming
out from the point you clicked (or '-' zooming out, away from the center
of the screen). A few other things in DetailedWorldMap were tweaked as
well.