Add Isaac32RNG (variant on IsaacRNG)
IsaacRNG and Isaac32RNG both have setBlock methods now, to fill a
potentially-large array with longs or ints, respectively. Isaac32RNG can
also fill float arrays with floats in your choice of the range `[0,1)`
or `[-1,1)`.

MasonNoise is currently a WIP and may lead to some benefit for
higher-dimensional noise (4D and 6D mainly).