Subtle, important 4D/6D SeededNoise tweaks
These reduce the appearance of gray gaps between very light or very dark
areas. Before, you could sometimes have large swaths of gray dominating
an otherwise colorful image, and now the surflets (small regions that
are more flat in color) are the emphasis instead of the gray. This is
the same as what 2D and 3D SeededNoise already had working, but the
parameters Joise used weren't quite right for our case.