Various cleanup in WMG
The zoomIn and zoomOut methods now take an amount to zoom by; this needs
better testing but should behave the same if the zoomAmount is 1. The
SimpleBiomeMapper inner class is simpler now. The heightCodeData field
is public now, as it should be; a few other fields also became more
accessible.