Key improvements to rivers in WMG
Rivers now should reach the coast much more often without running in
weird shapes around the coastline (not always in a contiguous way
before). There are also changes to how diagonals are drawn for rivers so
it will produce a solid orthogonal path more often than before
(tributaries almost always branch on a diagonal, though this is
intentional). The volume of rivers has been tweaked, but this may become
a parameter soon.

This was tested in a related project ( tommyettinger/ColorGuard ) before
adding it back to SquidLib. I'm pretty sure I put everything back where
it should be; if not, I should be able to tell as soon as I update the
snapshot...