Attempts at trigonometry-based noise functions
None have really been a success so far, but I also added some
potentially-useful methods to NumberTools (including the GWT emulated
version). The general idea for using trig functions as part of noise is
that they are already continuous and use floating-point positions, so
there may be some way to align, rotate, skew, or otherwise use the
periodic waves of sine or cosine functions to make a
less-noticeably-repetitive pattern.