Very careful tweaking of 6D SeededNoise
This continues the work from earlier to reduce the gray empty space
between surflets, which was only a problem in 4D and 6D and had recently
been fixed for 4D (where it is used more). 6D had received some changes,
but they weren't quite enough. I understand the NumberTools.bounce()
method better now, which is used to help limit the output to a
guaranteed range of -1.0 to 1.0, and this allowed me to experiment with
a wider range of permissiveness options for 6D noise. The results are
good! Before (at a frequent bad moment): http://i.imgur.com/pNhURLk.png
After (at a good moment, but even the bad moments aren't nearly as
noticeable): http://i.imgur.com/yvqGRz7.png