Refactor worldmap demo to use main code
As I had hoped, it is a lot simpler to use now, especially the zoom
code. Rivers have changed, again. The river start calculation now
doesn't consider the volume of the map's land as part of the fraction of
cells to mark as river starts, which had caused map size to
disproportionately affect river count. Rivers also extend a tiny bit out
to sea again, passing into the shallows like most real rivers do, though
like real rivers there's no indication that the area is a mix of salt
and fresh water (it mostly serves to guarantee rivers complete the trip
to the sea even when zoomed in).