API Change in WMG; can choose detail level
The previous API of WorldMapGenerator is all available in an inner
class, WorldMapGenerator.TilingMap, but now this allows the possibility
of other types of map generation based on different techniques, and not
only ones that seamlessly tile on both 2D axes (north-pole-to-south-pole
wrapping isn't necessarily optimal). Also, now you can give TilingMap a
multiplier for the octaves to use. 0.75 seems fine, even with a little
zooming, in the demo that uses this (and is a fair amount faster than
the default 1.0).