Major cleanup/improvement for docs.
As many of the documentation warnings as I could have been fixed. The
kind caused by a return tag having no parameter is too numerous to fix,
so those are still there. There's another warning caused by an actual
known bug in JDK 8's JavaDoc tool,
https://bugs.openjdk.java.net/browse/JDK-8031625 , that relates to inner
and nested classes being impossible to adequately document. Great job,
Oracle! You are doing so good at not breaking backwards compatibility in
APIs that you (think you) deserve to be able to but breaking changes in
any peripheral tool imaginable, like the Java 8 doclint debacle!

There's also some improvements to WorldMapGenerator, using less calls to
trig functions and behaving correctly when the map is as it should be --
about 3 times as wide as it is tall, for geometry reasons.