Major changes in squidlib-extra
The old Json-based API is still intact in the squidpony package of
-extra, but it delegates to the squidpony.store.json package. In the
future, it may be swappable between that and the new
squidpony.store.text package (which is just the TextStorage class and an
extension of Converters, BonusConverters), which seems to be even
smaller for saved data (about 10% better compressed, probably comparable
uncompressed) but needs a little extra work to use. On the plus side,
the techniques in StringConvert, which TextStorage uses heavily, are
easier to write for your own classes than the Json-based approach.
Several classes from Converters have received bugfixes. There's a minor
behavior-breaking change in LightRNG to improve randomness quality,
which only affects the constructor that takes a String, and a way to get
the exact behavior from before is mentioned in the docstring for that
constructor. BonusConverters in squidlib-extra can serialize StatefulRNG
but not RNG (yet). TextStorageTest shows that this all works correctly.
Phew...