Add BundleBiMap to handle BiMap/MultiMap things
This should be useful for the original purpose, as a way of handling
ingredients and quantities for a recipe/crafting system, across various
games, but it has some limitations currently that would be difficult for
uses other than rules like crafting (e.g. no deletion from the data
structure). I hope you like it @SquidPony !