Some likely fixes to MultiArrangement; needs tests
I'm not sure if I did all of these changes correctly, but now the size
of the hash table and the size of the iteration order are not
necessarily equal (we can have a key show up more than once in iteration
but only once in the hash table). I think I cleaned up most of the
incorrect usage of size when order.size was needed. KeySet is now
KeyList, as well, reflecting the non-unique keys.