Tweaks to ProbabilityTable, (Bonus)Converters
BonusConverters in squidlib-extra should be able to serialize a
ProbabilityTable now, though as usual it needs testing. ProbabilityTable
doesn't use a wildcard generic type for the extra tables, because the
information that a subclass adds to the main generic type probably
wouldn't be serializable. The code for the extra tables is simpler now
because we don't need identity hashing to get it to work.

There's some bugfixes to the bin() methods in StringKit for less-common
argument types. Coord3D has a correct toString method now, matching the
current style used by Coord.

MasonNoise still doesn't work.