Speedup on MasonNoise by unusual technique
MasonNoise is a kind of value noise, which normally would require
getting a hash or other random value per corner of a cell, in such a way
that shared corners share a random value. BUT, due to how this class
works, it only needs random values per axis, and can quickly mix the two
axes to get a random corner without hashing. There's a method in
MasonNoise now, addNoiseField(), that adds float values into a float[][]
but only calculates a value for a point on the the x-axis or y-axis
once, saving a significant amount of time. It also skews the grid based
on the seed you give, helping reduce right-angle bias when multiple
octaves are added together. With 4 octaves, it looks like
http://i.imgur.com/GWnqEc3.png .