Small and less-small fix to Radius.perimeter()
The perimeter method was only returning one Coord for circle radii, due
to a stupid mistake where a logical not was missing (one character to
fix). The less-small fix was related to an unusual case where a
two-cell-thick perimeter occurred infrequently; this seems to be
resolved by flooring towards 0 instead of rounding and by using a
slightly larger radius in some calculations.