Add 3D MasonNoise, GWT fix in NumberTools
In NumberTools in the main code, I had added a method but forgotten to
add an equivalent in the GWT super-sourced NumberTools (which should be
faster on GWT than just calling the original Java code). MasonNoise only
has 3D via the addNoiseField method, but both the 2D and 3D overloads of
that method have changed in this commit. They seem to be smoother, using
a variant on cubic interpolation that seems to be both faster and have a
better curve shape while retaining quintic interpolation's zeroing-in at
the start and end. 2D changed a lot yesterday while I was working on it,
and the results are good. 3D is extremely similar to 2D and this should
extend the advantages 2D MasonNoise has regarding pre-calculating large
areas.