Things for mixing images with text.
ImageSquidPanel is just like a regular SquidPanel (it's a subclass),
except that you can modify its imageMap field or call setImageSwap(), or
a few other methods, to have it replace a char with a TextureRegion
whenever the char would be drawn. Other methods have been modified in
that class to make this work smoothly. There's also new methods in
TextCellFactory to round out support for TextureRegions where it was
missing before, and also similar tweaks in SquidPanel. Two unused
protected methods in SquidPanel were removed because they were
unnecessary and somewhat confusing as to what was actually relevant.