Quick fix to SquidLayers layer setting; new demo
ImageLayersTest is a modified version of IconsTest to show images being
overlaid on top of both a solid background and a background image.
SquidLayers, in the last commit, didn't correctly update its Group
contents to match the foreground, background, or extra panels that were
set with setForegroundLayer and so on. Now it does set them. The demo
uses this by setting the foreground layer to an ImageSquidPanel, and it
looks like http://i.imgur.com/fBJfwx4.gifv .