Good improvements to Flap; Dash needs work
Dash had serious issues where it would enter low-period cycles from
certain seeds very quickly. Dash has been changed, but is slower. This
also means some of the randomizing methods that copied from Dash needed
to be changed (to the best one we have for int-to-random-int,
PintRNG.determine() ). FlapRNG is now based on XoRoRNG, but using two
ints for state instead of two longs. This is advantageous on GWT, but
also gives FlapRNG an edge over PintRNG when generating longs with
nextLong() (as RNG often requires), since it has enough bits of state to
cover a random long fairly well.