Keep on fixing! Arrangement.alter() works now
Even though it stopped changing the size of the Arrangement in the last
commit, the values were still becoming mis-aligned with their keys. This
should be fixed, and is tested much more thoroughly. OrderedMap was also
tested but showed no signs of a problem in the first place. OrderedSet
doesn't have values, so it can't have this problem.