Teensy fix resolves issues with Arrangement
Specifically, first and last can be set in methods called by
alterEntry() (which is used by alter() ), but in a way that isn't
reliable for that method unless the actual first/last status is checked
at the very start of the call. Now it should treat the first and last
elements correctly, just like the rest of the Arrangement.