CONTRIBUTORS, PLEASE CLONE AGAIN
It wasn't stored in a commit message, but between this commit and the
previous commit, the obsolete and very large-to-download change history
of the javadocs folder "docs/" and any large files previously in the
repo (that weren't also present in the current revision) were removed.
That means that if you send a pull request without cloning the repo
anew, then the old docs history may come back in pieces and flood the
repo again. The reason for doing this was to try to resolve a JitPack
issue that I suspected was related to the download size getting
unmanageable, and if we get the history returned it will go back to
having that same issue.

Also in this commit are some improvements to LightRNG.determine() (a
rarely-used static method) and CosmicNumbering (allowing it to more
often generate numbers in the full 0.0-1.0 range).