Some adjustments to ThinDungeonGen
You can now remove corners from the known dungeon in a
ThinDungeonGenerator, which has various effects on FOV and pathfinding.
This is usually beneficial when the FOV type is relatively strict,
because corners can block line of sight to areas that the player would
expect to be able to see. It does allow the player to "look inside" some
walls without special handling, however, because corners are used to
block light from going past the first thin boundary; this isn't a severe
problem because it doesn't permit the player any mechanical advantages
beyond knowing that a wall isn't single-cell-thin, and that's known
anyway if they would be looking into a corner (single-cell-wide lines
don't have corners).