Modify seed setup for LapRNG, fix next() in places
NLFSR and LapRNG both gave too many bits in their next() method (and
LapRNG would give lower-quality bits as well). This is fixed. LapRNG
turns out to be sensitive to one of its seeds' starting values (seed0 or
state0); if this number is smaller than 2 to the 25 or so, then the
sequence it produces will be very similar to other nearby numbers for
that seed. This is a problem in HashVisualizer because it uses a counter
as the seed, changing by 1 each time. To compensate for this, passing
one or two ints to LapRNG's constructor will multiply each by a large
constant and add another large constant. The same is done for the
one-long constructor.
The state variables in LapRNG are now private with setters and getters,
so at least documenting the bad settings can be done more reasonably.